+++ Project/Source/Cells/Helpers+Additions/NSStringAdditions.h
@@ -38,12 +38,12 @@
 + (NSString *) stringForSpeed: (CGFloat) speed;
 + (NSString *) stringForSpeedAbbrev: (CGFloat) speed;
 + (NSString *) stringForRatio: (CGFloat) ratio;
-+ (NSString *) stringForTime: (CGFloat) seconds;
++ (NSString *) stringForTime: (int) seconds;
 
 + (NSString *) percentString: (CGFloat) progress longDecimals: (BOOL) longDecimals;
 
-+ (NSString *) timeString: (uint64_t) seconds showSeconds: (BOOL) showSeconds;
-+ (NSString *) timeString: (uint64_t) seconds showSeconds: (BOOL) showSeconds maxFields: (NSUInteger) max;
++ (NSString *) timeString: (int) seconds showSeconds: (BOOL) showSeconds;
++ (NSString *) timeString: (int) seconds showSeconds: (BOOL) showSeconds maxFields: (NSUInteger) max;
 
 - (NSComparisonResult) compareNumeric: (NSString *) string; //simple compare method for strings with numbers (works for IP addresses)
